refactor(finance): change currency data type from double to long

+++ client/src/main/java/gui/Transaction.java
@@ -1,19 +1,19 @@
 package gui;
 
 import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleLongProperty;
 
 public class Transaction {
     private final SimpleStringProperty date;
     private final SimpleStringProperty type;
-    private final SimpleDoubleProperty amount;
+    private final SimpleLongProperty amount;
     private final SimpleStringProperty status;
     private final SimpleStringProperty note;
 
-    public Transaction(String date, String type, double amount, String status, String note) {
+    public Transaction(String date, String type, long amount, String status, String note) {
         this.date = new SimpleStringProperty(date);
         this.type = new SimpleStringProperty(type);
-        this.amount = new SimpleDoubleProperty(amount);
+        this.amount = new SimpleLongProperty(amount);
         this.status = new SimpleStringProperty(status);
         this.note = new SimpleStringProperty(note);
     }
@@ -24,7 +24,7 @@ public class Transaction {
     public String getType() {
         return type.get();
     }
-    public double getAmount() {
+    public long getAmount() {
         return amount.get();
     }
     public String getStatus() {
